Fix type argument.

+++ cpp/subprojects/common/include/common/prediction/predictor_probability.hpp
@@ -15,4 +15,4 @@ typedef IPredictor<DensePredictionMatrix<float64>> IProbabilityPredictor;
 /**
  * Defines an interface for all classes that allow to create instances of the type `IProbabilityPredictor`.
  */
-typedef IPredictorFactory<DensePredictionMatrix<uint8>> IProbabilityPredictorFactory;
+typedef IPredictorFactory<DensePredictionMatrix<float64>> IProbabilityPredictorFactory;
